requirements.txt
distro==1.8.0
greenlet==2.0.2
numpy==1.24.2
pandas==1.5.3
pyodbc==4.0.35
pyserial==3.5
PySimpleGUI==4.60.4
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.5.post1
tabula-py==2.6.0
typing_extensions==4.5.0
